Refactor code and close window on bye command
