+++ src/main/java/mnsky/Launcher.java
@@ -1,6 +1,7 @@
 package mnsky;
 
 import javafx.application.Application;
+import mnsky.gui.Main;
 
 public class Launcher {
     public static void main(String[] args) {
