+++ src/main/java/mnsky/core/TaskList.java
@@ -1,4 +1,4 @@
-package mnsky;
+package mnsky.core;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
