+++ src/main/java/mnsky/core/Ui.java
@@ -1,4 +1,4 @@
-package mnsky;
+package mnsky.core;
 
 import java.util.ArrayList;
