+++ src/main/java/mnsky/gui/DialogBox.java
@@ -1,4 +1,4 @@
-package mnsky;
+package mnsky.gui;
 
 import java.io.IOException;
 import java.util.Collections;
