+++ src/main/java/mnsky/gui/Main.java
@@ -1,4 +1,4 @@
-package mnsky;
+package mnsky.gui;
 
 import java.io.IOException;
 
@@ -7,6 +7,7 @@ import javafx.fxml.FXMLLoader;
 import javafx.scene.Scene;
 import javafx.scene.layout.AnchorPane;
 import javafx.stage.Stage;
+import mnsky.core.Mnsky;
 
 /**
  * A GUI for Duke using FXML.
