+++ src/test/java/mnsky/ParserTest.java
@@ -5,6 +5,7 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import java.util.ArrayList;
 import java.util.List;
 
+import mnsky.core.Parser;
 import org.junit.jupiter.api.Test;
 
 public class ParserTest {
