+++ src/test/java/mnsky/TaskListTest.java
@@ -4,6 +4,9 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import java.util.ArrayList;
 
+import mnsky.core.Storage;
+import mnsky.core.TaskList;
+import mnsky.core.Ui;
 import org.junit.jupiter.api.Test;
 
 import mnsky.task.Task;
